services/websocket-service/requirements.txt
python-socketio>=5.10.0
aiohttp>=3.9.0
aiokafka>=0.10.0
redis>=5.0.0
python-dotenv>=1.0.0
